requirement.txt
dotenv
pandasai
streamlit